Death & Revival added ep.19

+++ July Jam - Elden Ring/Assets/Scripts/Game Saving/CharacterSaveData.cs
@@ -23,7 +23,7 @@ public class CharacterSaveData
     public float zPos;
 
     [Header("Resources")]
-    public float currentHealth;
+    public int currentHealth;
     public float currentStamina;
 
     [Header("Stats")]
